requirements.txt
absl-py==2.1.0
aniso8601==9.0.1
args==0.1.0
asttokens==2.4.1
astunparse==1.6.3
attacut==1.0.6
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.8.2
botnoi==0.2.9
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clint==0.5.1
cmake==3.30.3
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
decorator==5.1.1
deepcut==0.7.0.0
dlib==19.24.6
docopt==0.6.2
editdistance==0.8.1
epitran==1.25
et-xmlfile==1.1.0
executing==2.0.1
face-Recog==0.0.1
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.14.0
fire==0.6.0
Flask==3.0.3
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.10
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.5.0
future==1.0.0
gast==0.5.4
gensim==4.3.3
google-pasta==0.2.0
grpcio==1.64.0
gunicorn==19.6.0
h5py==3.11.0
httplib2==0.22.0
idna==3.7
intel-openmp==2021.4.0
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
marisa-trie==1.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
munkres==1.1.4
namex==0.0.8
netmiko==4.3.0
networkx==3.3
nltk==3.8.1
nptyping==2.5.0
ntc_templates==4.3.0
numpy==1.26.4
oauth2client==4.1.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.3
panphon==0.20.0
paramiko==3.4.0
parso==0.8.4
pillow==10.4.0
prompt_toolkit==3.0.47
protobuf==4.25.3
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
Pygments==2.18.0
PyNaCl==1.5.0
pyparsing==3.1.2
pyserial==3.5
pythainlp==5.0.3
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
regex==2024.5.15
requests==2.32.2
rich==13.7.1
rq==1.16.2
rsa==4.9
scikit-learn==1.5.0
scp==0.14.5
setuptools==74.1.2
six==1.16.0
sklearn-crfsuite==0.3.6
smart-open==7.0.4
soupsieve==2.6
ssg==0.0.8
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tbb==2021.12.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
textfsm==1.1.3
threadpoolctl==3.5.0
timeout-decorator==0.5.0
torch==2.3.0
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.0
tzdata==2024.1
unicodecsv==0.14.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.44.0
wrapt==1.16.0
xlrd==2.0.1